# docs/api_specification.yaml
openapi: 3.0.3
info:
  title: バー検索アプリ API
  description: |
    バー検索アプリのバックエンドAPI仕様書

    ## 認証
    Supabase Authを使用した認証を実装。
    全ての保護されたエンドポイントには `Authorization: Bearer {token}` ヘッダーが必要。

    ## ベースURL
    - 開発環境: `http://localhost:8000`
    - 本番環境: `https://api.bar-search-app.com` (例)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: ローカル開発環境
  - url: https://api.bar-search-app.com
    description: 本番環境
...
paths:
...
  /api/users/me:
    get:
      tags:
        - users
      summary: 自分のプロフィール取得
      description: ログインユーザーのプロフィール情報を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: プロフィール取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - users
      summary: プロフィール更新
      description: ログインユーザーのプロフィール情報を更新
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: プロフィール更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: リクエストボディ不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/me/reviews:
    get:
      tags:
        - users
      summary: 自分が投稿したレビュー一覧取得
      description: ログインユーザーが投稿した全レビューを取得
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: 取得件数
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: オフセット
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: レビュー一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewWithBar'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: 未認証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# ========================================
# コンポーネント定義
# ========================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase AuthのJWTトークン
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
        details:
          type: object
          description: 詳細情報（オプション）

    # バー関連
...
    Review:
      type: object
      required:
        - id
        - bar_id
        - user_id
        - rating
      properties:
        id:
          type: string
          format: uuid
          description: レビューID
        bar_id:
          type: string
          format: uuid
          description: バーID
        user_id:
          type: string
          format: uuid
          description: ユーザーID
        user_nickname:
          type: string
          description: 投稿者のニックネーム
        user_avatar_url:
          type: string
          format: uri
          description: 投稿者のアイコンURL
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: 星評価（1-5）
        comment:
          type: string
          maxLength: 2000
          description: レビューコメント
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時
    ReviewWithBar:
      allOf:
        - $ref: '#/components/schemas/Review'
        - type: object
          properties:
            bar_name:
              type: string
              description: バー名
            bar_address:
              type: string
              description: バー住所
...
    UserProfile:
      type: object
      required:
        - user_id
        - email
        - nickname
        - age
      properties:
        user_id:
          type: string
          format: uuid
          description: ユーザーID
        email:
          type: string
          format: email
          description: メールアドレス
        nickname:
          type: string
          maxLength: 50
          description: ニックネーム
        age:
          type: integer
          minimum: 1
          maximum: 150
          description: 年齢
        avatar_url:
          type: string
          format: uri
          description: アイコン写真URL
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時
    UserProfileUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: メールアドレス
        nickname:
          type: string
          maxLength: 50
          description: ニックネーム
        age:
          type: integer
          minimum: 1
          maximum: 150
          description: 年齢
        avatar_url:
          type: string
          format: uri
          description: アイコン写真URL